basic

+++ libsrc/libs/memory.h
@@ -31,7 +31,7 @@ int sc_regInit(void);
 int sc_regSet(int register, int value);
 int sc_regGet(int register, int *value);
 
-int sc_commandEncode(int command, int operand, int *value);
+int sc_commandEncode(int command, int operand, short *value);
 int sc_commandDecode(int value, int *command, int *operand);
 
 #endif
